requirements.txt
nicegui
pandas
openpyxl
google-generativeai
python-dotenv
Pillow
plotly
requests
supabase
